+++ apx_sdk/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support SHA-512 content digests.
 - Support verification of signed requests with `Content-Digest` header.
 - Added `NoDigest` variant to `HttpSignatureVerificationError` enum.
+- Support verification of RFC-9421 signatures.
 
 ### Changed